+++ sdk/resourcemanager/sql/armsql/zz_generated_response_types.go
@@ -73,6 +73,21 @@ type DataWarehouseUserActivitiesClientListByDatabaseResponse struct {
 	DataWarehouseUserActivitiesListResult
 }
 
+// DatabaseAdvancedThreatProtectionSettingsClientCreateOrUpdateResponse contains the response from method DatabaseAdvancedThreatProtectionSettingsClient.CreateOrUpdate.
+type DatabaseAdvancedThreatProtectionSettingsClientCreateOrUpdateResponse struct {
+	DatabaseAdvancedThreatProtection
+}
+
+// DatabaseAdvancedThreatProtectionSettingsClientGetResponse contains the response from method DatabaseAdvancedThreatProtectionSettingsClient.Get.
+type DatabaseAdvancedThreatProtectionSettingsClientGetResponse struct {
+	DatabaseAdvancedThreatProtection
+}
+
+// DatabaseAdvancedThreatProtectionSettingsClientListByDatabaseResponse contains the response from method DatabaseAdvancedThreatProtectionSettingsClient.ListByDatabase.
+type DatabaseAdvancedThreatProtectionSettingsClientListByDatabaseResponse struct {
+	DatabaseAdvancedThreatProtectionListResult
+}
+
 // DatabaseAdvisorsClientGetResponse contains the response from method DatabaseAdvisorsClient.Get.
 type DatabaseAdvisorsClientGetResponse struct {
 	Advisor
@@ -1488,6 +1503,31 @@ type ManagedRestorableDroppedDatabaseBackupShortTermRetentionPoliciesClientUpdat
 	ManagedBackupShortTermRetentionPolicy
 }
 
+// ManagedServerDNSAliasesClientAcquireResponse contains the response from method ManagedServerDNSAliasesClient.Acquire.
+type ManagedServerDNSAliasesClientAcquireResponse struct {
+	ManagedServerDNSAlias
+}
+
+// ManagedServerDNSAliasesClientCreateOrUpdateResponse contains the response from method ManagedServerDNSAliasesClient.CreateOrUpdate.
+type ManagedServerDNSAliasesClientCreateOrUpdateResponse struct {
+	ManagedServerDNSAlias
+}
+
+// ManagedServerDNSAliasesClientDeleteResponse contains the response from method ManagedServerDNSAliasesClient.Delete.
+type ManagedServerDNSAliasesClientDeleteResponse struct {
+	// placeholder for future response values
+}
+
+// ManagedServerDNSAliasesClientGetResponse contains the response from method ManagedServerDNSAliasesClient.Get.
+type ManagedServerDNSAliasesClientGetResponse struct {
+	ManagedServerDNSAlias
+}
+
+// ManagedServerDNSAliasesClientListByManagedInstanceResponse contains the response from method ManagedServerDNSAliasesClient.ListByManagedInstance.
+type ManagedServerDNSAliasesClientListByManagedInstanceResponse struct {
+	ManagedServerDNSAliasListResult
+}
+
 // ManagedServerSecurityAlertPoliciesClientCreateOrUpdateResponse contains the response from method ManagedServerSecurityAlertPoliciesClient.CreateOrUpdate.
 type ManagedServerSecurityAlertPoliciesClientCreateOrUpdateResponse struct {
 	ManagedServerSecurityAlertPolicy
@@ -1691,6 +1731,21 @@ type SensitivityLabelsClientListRecommendedByDatabaseResponse struct {
 // SensitivityLabelsClientUpdateResponse contains the response from method SensitivityLabelsClient.Update.
 type SensitivityLabelsClientUpdateResponse struct {
 	// placeholder for future response values
+}
+
+// ServerAdvancedThreatProtectionSettingsClientCreateOrUpdateResponse contains the response from method ServerAdvancedThreatProtectionSettingsClient.CreateOrUpdate.
+type ServerAdvancedThreatProtectionSettingsClientCreateOrUpdateResponse struct {
+	ServerAdvancedThreatProtection
+}
+
+// ServerAdvancedThreatProtectionSettingsClientGetResponse contains the response from method ServerAdvancedThreatProtectionSettingsClient.Get.
+type ServerAdvancedThreatProtectionSettingsClientGetResponse struct {
+	ServerAdvancedThreatProtection
+}
+
+// ServerAdvancedThreatProtectionSettingsClientListByServerResponse contains the response from method ServerAdvancedThreatProtectionSettingsClient.ListByServer.
+type ServerAdvancedThreatProtectionSettingsClientListByServerResponse struct {
+	LogicalServerAdvancedThreatProtectionListResult
 }
 
 // ServerAdvisorsClientGetResponse contains the response from method ServerAdvisorsClient.Get.